regular update from dev

+++ script/get-mlperf-inference-loadgen/run.sh
@@ -36,7 +36,9 @@ if [ "${?}" != "0" ]; then exit 1; fi
 
 # Clean build directory (too large)
 cd "${CUR_DIR}"
-rm -rf build
+if [[ $CM_MLPERF_INFERENCE_LOADGEN_BUILD_CLEAN == "yes" ]]; then
+  rm -rf build
+fi
 
 PYTHON_VERSION=`${CM_PYTHON_BIN_WITH_PATH} -V |cut -d' ' -f2`
 PYTHON_SHORT_VERSION=${PYTHON_VERSION%.*}
@@ -44,12 +46,14 @@ PYTHON_MINOR_VERSION=${PYTHON_SHORT_VERSION#*.}
 MLPERF_INFERENCE_PYTHON_SITE_BASE=${INSTALL_DIR}"/python"
 
 cd "${CM_MLPERF_INFERENCE_SOURCE}/loadgen"
-CFLAGS="-std=c++14 -O3" ${CM_PYTHON_BIN_WITH_PATH} setup.py bdist_wheel
-${CM_PYTHON_BIN_WITH_PATH} -m pip install --force-reinstall `ls dist/mlperf_loadgen-*cp3${PYTHON_MINOR_VERSION}*.whl` --target=${MLPERF_INFERENCE_PYTHON_SITE_BASE}
+#CFLAGS="-std=c++14 -O3" ${CM_PYTHON_BIN_WITH_PATH} setup.py bdist_wheel
+#${CM_PYTHON_BIN_WITH_PATH} -m pip install --force-reinstall `ls dist/mlperf_loadgen-*cp3${PYTHON_MINOR_VERSION}*.whl` --target="${MLPERF_INFERENCE_PYTHON_SITE_BASE}"
+${CM_PYTHON_BIN_WITH_PATH} -m pip install .
+
 if [ "${?}" != "0" ]; then exit 1; fi
 
 # Clean the built wheel
-find . -name 'mlperf_loadgen*.whl' | xargs rm
+find . -name 'mlcommons_loadgen*.whl' | xargs rm
 
 echo "******************************************************"
 echo "Loadgen is built and installed to ${INSTALL_DIR} ..."
